+++ cocos/core/root.ts
@@ -5,8 +5,9 @@
 import { builtinResMgr } from './3d/builtin';
 import { GFXDevice } from './gfx/device';
 import { Pool } from './memop';
-import { RenderPipeline } from './pipeline/render-pipeline';
-import { IRenderViewInfo, RenderView } from './pipeline/render-view';
+import { RenderPipeline } from './pipeline';
+import { RenderView } from './pipeline';
+import { IRenderViewInfo } from './pipeline/define'
 import { Camera, Light, Model } from './renderer';
 import { DataPoolManager } from './renderer/data-pool-manager';
 import { LightType } from './renderer/scene/light';
@@ -16,7 +17,7 @@ import { SpotLight } from './renderer/scene/spot-light';
 import { UI } from './renderer/ui/ui';
 import { legacyCC } from './global-exports';
 import { RenderWindow, IRenderWindowInfo } from './renderer/core/render-window';
-import { ForwardPipeline } from './pipeline/forward/forward-pipeline';
+import { ForwardPipeline } from './pipeline';
 import { GFXColorAttachment, GFXDepthStencilAttachment, GFXStoreOp } from './gfx';
 import { RootHandle, RootPool, RootView, NULL_HANDLE } from './renderer/core/memory-pools';
